Room component does most of the work now weeeeeeeeeeee
