+++ Assets/Scripts/Oxygen/RepairZone.cs
@@ -5,12 +5,18 @@ using UnityEngine;
 public class RepairZone : MonoBehaviour
 {
     public bool isOccupied { get; set; }
+    public PLAYER occupyingPlayer = PLAYER.NONE;
 
     void OnTriggerStay2D(Collider2D col)
     {
         if(col.tag == "Player")
         {
             isOccupied = true;
+
+            if (occupyingPlayer == PLAYER.NONE)
+            {
+                //occupyingPlayer = col.gameObject.GetComponent<PlayerScript>().player;
+            }
         }
 
     }
@@ -19,6 +25,7 @@ public class RepairZone : MonoBehaviour
         if (col.tag == "Player")
         {
             isOccupied = false;
+            occupyingPlayer = PLAYER.NONE;
         }
     }
 }
