+++ Assets/Scripts/PlayerEnums.cs
@@ -1,0 +1,6 @@
+﻿public enum PLAYER
+{
+    PLAYER1,
+    PLAYER2,
+    NONE
+};
