+++ Assets/Scripts/PlayerEnums.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ec3623136e2bdb4ea1fc8931ac7020a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
